separated override table rows into separate components to make it easier to manage the state of the technology, override values; added functionality to incorporate value names for other infrastructure buildout types
